fix circular dependency and address Robert's comments
