+++ pact/Pact/Core/Environment/Types.hs
@@ -46,8 +46,6 @@ module Pact.Core.Environment.Types
  , defaultEvalEnv
  , GasLogEntry(..)
  , RecursionCheck(..)
- , Bytes
- , SizeOfByteLimit(..)
  ) where
 
 
@@ -60,7 +58,6 @@ import Data.Map.Strict(Map)
 import Data.List.NonEmpty (NonEmpty(..))
 import Data.IORef
 import Data.Default
-import Data.Word (Word64)
 
 import Control.DeepSeq
 import GHC.Generics
@@ -70,7 +67,6 @@ import qualified Data.Text as T
 import qualified Data.Map.Strict as M
 
 import Pact.Core.Persistence.Types
-import Pact.Core.Pretty
 import Pact.Core.Capabilities
 import Pact.Core.Guards
 import Pact.Core.PactValue
@@ -81,7 +77,6 @@ import Pact.Core.ChainData
 import Pact.Core.Errors
 import Pact.Core.Gas
 import Pact.Core.Namespace
-import Pact.Core.SizeOf
 import Pact.Core.StackFrame
 import Pact.Core.Builtin (IsBuiltin)
 import Pact.Core.SPV
@@ -191,7 +186,7 @@ data EvalState b i
 instance (NFData b, NFData i) => NFData (EvalState b i)
 
 instance Default (EvalState b i) where
-  def = EvalState def [] [] mempty Nothing Nothing (RecursionCheck mempty :| []) 0
+  def = EvalState def [] [] mempty Nothing Nothing (RecursionCheck mempty :| [])
 
 makeClassy ''EvalState
 
@@ -239,12 +234,3 @@ defaultEvalEnv pdb m = do
     , _eeGasModel = freeGasModel
     , _eeSPVSupport = noSPVSupport
     }
-
-type Bytes = Word64
-
-newtype SizeOfByteLimit
-  = SizeOfByteLimit Bytes
-  deriving Show
-
-instance Pretty SizeOfByteLimit where
-  pretty = pretty . show
